Add files via upload
the Main.py file contains the code for actual processing. While the functions.py contains the code for the functions the jarpy performs. The speaker.py contains the code for the output in the voice format. Remember the it is for macOS and need to be modified for windows. and the test.py contains the code and it provides platform to test new functions without actually adding them to the functions.py. Happy coding...!